Working on WOLF3D-Extractor port

+++ Graphics/PaletteFileReader.cs
@@ -6,9 +6,9 @@ namespace WOLF3D.Graphics
     {
         private static readonly int COLORS = 256;
 
-        public static int[] ColorModelFromPAL(string file)
+        public static uint[] ColorModelFromPAL(string file)
         {
-            int[] result = new int[COLORS];
+            uint[] result = new uint[COLORS];
             using (StreamReader input = new StreamReader(file))
             {
                 if (!input.ReadLine().Equals("JASC-PAL") || !input.ReadLine().Equals("0100"))
@@ -24,7 +24,10 @@ namespace WOLF3D.Graphics
                         || !byte.TryParse(tokens[1], out byte g)
                         || !byte.TryParse(tokens[2], out byte b))
                         throw new InvalidDataException("Palette \"" + file + "\" is an incorrectly formatted JASC palette.");
-                    result[x] = (r << 24) + (g << 16) + (b << 8) + (x == 255 ? 0 : 255);
+                    result[x] = (uint)(r << 24) 
+                        + (uint)(g << 16) 
+                        + (uint)(b << 8) 
+                        + (uint)(x == 255 ? 0 : 255);
                 }
             }
             return result;
